Simplify File analysis/validation
