+++ src/Altinn.App.Core/Extensions/ServiceCollectionExtensions.cs
@@ -4,7 +4,6 @@ using Altinn.App.Core.Features;
 using Altinn.App.Core.Features.Action;
 using Altinn.App.Core.Features.DataLists;
 using Altinn.App.Core.Features.DataProcessing;
-using Altinn.App.Core.Features.FileAnalyzis;
 using Altinn.App.Core.Features.Options;
 using Altinn.App.Core.Features.PageOrder;
 using Altinn.App.Core.Features.Pdf;
@@ -163,8 +162,6 @@ namespace Altinn.App.Core.Extensions
             AddPdfServices(services);
             AddEventServices(services);
             AddProcessServices(services);
-            AddFileAnalyserServices(services);
-            AddFileValidatorServices(services);
             AddMetricsDecorators(services, configuration);
 
             if (!env.IsDevelopment())
@@ -272,18 +269,6 @@ namespace Altinn.App.Core.Extensions
             services.AddTransient<IUserAction, SigningUserAction>();
             services.AddHttpClient<ISignClient, SignClient>();
             services.AddTransientUserActionAuthorizerForActionInAllTasks<UniqueSignatureAuthorizer>("sign");
-        }
-
-        private static void AddFileAnalyserServices(IServiceCollection services)
-        {
-            services.TryAddTransient<IFileAnalysisService, FileAnalysisService>();
-            services.TryAddTransient<IFileAnalyserFactory, FileAnalyserFactory>();
-        }
-
-        private static void AddFileValidatorServices(IServiceCollection services)
-        {
-            services.TryAddTransient<IFileValidationService, FileValidationService>();
-            services.TryAddTransient<IFileValidatorFactory, FileValidatorFactory>();
         }
 
         private static void AddMetricsDecorators(IServiceCollection services, IConfiguration configuration)
